+++ packages/mermaid/src/diagrams/info/infoDb.ts
@@ -1,7 +1,9 @@
 import type { InfoFields, InfoDB } from './infoTypes.js';
-import { version } from '../../../package.json';
+import packageJson from '../../../package.json' assert { type: 'json' };
 
-export const DEFAULT_INFO_DB: InfoFields = { version } as const;
+export const DEFAULT_INFO_DB: InfoFields = {
+  version: packageJson.version + (includeLargeFeatures ? '' : '-tiny'),
+} as const;
 
 export const getVersion = (): string => DEFAULT_INFO_DB.version;